+++ test/cases/optimization-imported-css-inline-source-map/expected/main.js
@@ -1,12 +1,12 @@
 /******/ (() => { // webpackBootstrap
 /******/ 	"use strict";
 
-;// CONCATENATED MODULE: ./src/component.js
+;// ./src/component.js
 
 
 console.log('>> component');
 
-;// CONCATENATED MODULE: ./src/main.js
+;// ./src/main.js